refactor(lower component): refactor to remove lower component functions into home view
